Fix mobile email rendering, drop all paddings on vertical screens (#20)
The style-tag was scrapping the styles block, which was misleading. Template blocks should probably wrap DOM not CSS, especially with the naming.

I also used a zero-width space, or zero byte char, instead of the space, to enable paddings on Outlook.

The vertical padding is gone, on vertical phones, since most email clients implement their padding. We cannot drop the padding completely thanks to Mail on iOS, but we increased it to match iOS and look more deliberate in Gmail.